+++ insteon_mqtt/mqtt/KeypadLinc.py
@@ -5,22 +5,21 @@
 #===========================================================================
 import functools
 from .. import log
-from .Dimmer import Dimmer
 from .MsgTemplate import MsgTemplate
 
 LOG = log.get_logger()
 
 
-class KeypadLinc(Dimmer):
+class KeypadLinc:
     """TODO: doc
     """
     def __init__(self, mqtt, device):
         """TODO: doc
         """
-        # Dimmer will handle signals to/from the dimmer load button.
-        super().__init__(mqtt, device)
+        self.mqtt = mqtt
+        self.device = device
 
-        # Output state change reporting template.
+        # Output on/off state change reporting template.
         self.msg_btn_state = MsgTemplate(
             topic='insteon/{{address}}/state/{{button}}',
             payload='{{on_str.lower()}}',
@@ -38,7 +37,24 @@ class KeypadLinc(Dimmer):
             payload='{ "cmd" : "{{value.lower()}}" }',
             )
 
-        device.signal_pressed.connect(self.handle_pressed)
+        self.msg_dimmer_state = None
+        self.msg_dimmer_level = None
+        if self.device.is_dimmer:
+            # Output dimmer state change reporting template.
+            self.msg_dimmer_state = MsgTemplate(
+                topic='insteon/{{address}}/state/1',
+                payload='{ "state" : "{{on_str.upper()}}", '
+                        '"brightness" : {{level_255}} }',
+                )
+
+            # Input dimmer level command template.
+            self.msg_dimmer_level = MsgTemplate(
+                topic='insteon/{{address}}/level',
+                payload='{ "cmd" : "{{json.state.lower()}}", '
+                        '"level" : {{json.brightness}} }',
+                )
+
+        device.signal_active.connect(self.handle_active)
 
     #-----------------------------------------------------------------------
     def load_config(self, config, qos=None):
@@ -49,27 +65,22 @@ class KeypadLinc(Dimmer):
                     config is stored in config['keypad_linc'].
           qos:      The default quality of service level to use.
         """
-        # Load the dimmer configuration from the dimmer area, not the
-        # fanlinc area.
-        super().load_config(config, qos)
-
-        # Now load the fan control configuration
         data = config.get("keypad_linc", None)
-        self.load_keypad_config(data, qos)
-
-    #-----------------------------------------------------------------------
-    def load_keypad_config(self, config, qos=None):
-        """TODO: doc
-        """
-        if not config:
+        if not data:
             return
 
-        self.msg_btn_state.load_config(config, 'btn_state_topic',
+        self.msg_btn_state.load_config(data, 'btn_state_topic',
                                        'btn_state_payload', qos)
-        self.msg_btn_on_off.load_config(config, 'btn_on_off_topic',
+        self.msg_btn_on_off.load_config(data, 'btn_on_off_topic',
                                         'btn_on_off_payload', qos)
-        self.msg_btn_scene.load_config(config, 'btn_scene_topic',
+        self.msg_btn_scene.load_config(data, 'btn_scene_topic',
                                        'btn_scene_payload', qos)
+
+        if self.device.is_dimmer:
+            self.msg_dimmer_state.load_config(data, 'dimmer_state_topic',
+                                              'dimmer_state_payload', qos)
+            self.msg_dimmer_level.load_config(data, 'dimmer_level_topic',
+                                              'dimmer_level_payload', qos)
 
     #-----------------------------------------------------------------------
     def subscribe(self, link, qos):
@@ -79,8 +90,6 @@ class KeypadLinc(Dimmer):
           link:   The MQTT network client to use.
           qos:    The quality of service to use.
         """
-        super().subscribe(link, qos)
-
         # We need to subscribe to each button topic so we know which one is
         # which.
         for group in range(1, 9):
@@ -94,6 +103,12 @@ class KeypadLinc(Dimmer):
             topic = self.msg_btn_scene.render_topic(data)
             link.subscribe(topic, qos, handler)
 
+        # Add the input dimmer topic if needed.
+        if self.device.is_dimmer:
+            data = self.template_data(button=1)
+            topic = self.msg_dimmer_level.render_topic(data)
+            link.subscribe(topic, qos, self.handle_set_level)
+
     #-----------------------------------------------------------------------
     def unsubscribe(self, link):
         """Unsubscribe to any MQTT topics the object was subscribed to.
@@ -101,8 +116,6 @@ class KeypadLinc(Dimmer):
         Args:
           link:   The MQTT network client to use.
         """
-        super().unsubscribe(link)
-
         for group in range(1, 9):
             data = self.template_data(button=group)
 
@@ -111,6 +124,11 @@ class KeypadLinc(Dimmer):
 
             topic = self.msg_btn_scene.render_topic(data)
             link.unsubscribe(topic)
+
+        if self.device.is_dimmer:
+            data = self.template_data(button=1)
+            topic = self.msg_dimmer_level.render_topic(data)
+            self.mqtt.unsubscribe(topic)
 
     #-----------------------------------------------------------------------
     # pylint: disable=arguments-differ
@@ -159,6 +177,34 @@ class KeypadLinc(Dimmer):
             return
 
     #-----------------------------------------------------------------------
+    def handle_set_level(self, client, data, message):
+        """TODO: doc
+        """
+        LOG.info("KeypadLinc message %s %s", message.topic, message.payload)
+        assert self.msg_dimmer_level is not None
+
+        data = self.msg_dimmer_level.to_json(message.payload)
+        if not data:
+            return
+
+        LOG.info("KeypadLinc input command: %s", data)
+        try:
+            cmd = data.get('cmd')
+            if cmd == 'on':
+                level = int(data.get('level'))
+            elif cmd == 'off':
+                level = 0
+            else:
+                raise Exception("Invalid dimmer cmd input '%s'" % cmd)
+
+            instant = bool(data.get('instant', False))
+        except:
+            LOG.exception("Invalid dimmer command: %s", data)
+            return
+
+        self.device.set(level=level, instant=instant)
+
+    #-----------------------------------------------------------------------
     def handle_scene(self, client, data, message, group):
         """TODO: doc
         """
@@ -166,7 +212,7 @@ class KeypadLinc(Dimmer):
                   message.payload)
 
         # Parse the input MQTT message.
-        data = self.msg_scene_on_off.to_json(message.payload)
+        data = self.msg_btn_scene.to_json(message.payload)
         if not data:
             return
 
@@ -187,7 +233,7 @@ class KeypadLinc(Dimmer):
         self.device.scene(is_on, group)
 
     #-----------------------------------------------------------------------
-    def handle_pressed(self, device, group, is_active):
+    def handle_active(self, device, group, level):
         """Device active button pressed callback.
 
         This is triggered via signal when the Insteon device button is
@@ -197,12 +243,16 @@ class KeypadLinc(Dimmer):
         Args:
           device:   (device.Base) The Insteon device that changed.
           group:    (int) The button number 1...n that was pressed.
+          level:    (int) The current device level 0...0xff.
         """
-        LOG.info("MQTT received button press %s = btn %s on %s", device.label,
-                 group, is_active)
+        LOG.info("MQTT received button press %s = btn %s at %s", device.label,
+                 group, level)
 
-        level = 0x00 if not is_active else 0xff
         data = self.template_data(level, group)
-        self.msg_btn_state.publish(self.mqtt, data)
+
+        if group == 1 and self.device.is_dimmer:
+            self.msg_dimmer_state.publish(self.mqtt, data)
+        else:
+            self.msg_btn_state.publish(self.mqtt, data)
 
     #-----------------------------------------------------------------------